* bug d'arrondi dans les structural ctrl
* readaptation des proba des sensors structural

* mise a jour des éléments des listes de fonctions JS

* sensors et effectors dans un nouvel onglet pour les mutations

git-svn-id: http://www.genecraft-project.com/svn/genecraft/trunk@368 2ce2bb41-2ff2-4fd5-96a4-e3c75a7bd90e

+++ src/genelab/core/entities/mutation/structuralmutation.cpp
@@ -19,9 +19,9 @@ StructuralMutation::StructuralMutation(QVariant variant)
     QVariantMap map = variant.toMap();
 
     enable              = map["enable"].toBool();
-    addProbability      = map["addProbability"].toFloat();
-    deleteProbability   = map["deleteProbability"].toFloat();
-    replaceProbability  = map["replaceProbability"].toFloat();
+    addProbability      = map["addProbability"].toDouble();
+    deleteProbability   = map["deleteProbability"].toDouble();
+    replaceProbability  = map["replaceProbability"].toDouble();
 }
 
 bool StructuralMutation::checkAdd()
